+++ libraries/ghc-bignum/Dummy.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+module Dummy where
+
+import GHC.Internal.Bignum.Integer ()
+  -- See Note [Tracking dependencies on primitives] in GHC.Internal.Base