+++ libraries/ghc-internal/src/GHC/Internal/Integer.hs
@@ -57,8 +57,8 @@ module GHC.Internal.Integer (
     hashInteger,
     ) where
 
-import GHC.Num.Integer (Integer)
-import qualified GHC.Num.Integer as I
+import GHC.Internal.Bignum.Integer (Integer)
+import qualified GHC.Internal.Bignum.Integer as I
 import GHC.Prim
 import GHC.Types